python/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
elevenlabs==0.2.27  # Downgraded from 1.52.0 to 0.2.27
fonttools==4.56.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
noisereduce==3.0.3
numpy==2.2.3
openai==1.50.0  # Downgraded from 1.51.0 to 1.50.0
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0
python-dotenv==1.0.1  # Add this line